Combined microservice view helpers
