+++ Source/SelfService/Web/microservice/helpers/dateHelpers.tsx
@@ -1,17 +1,8 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-import React, { useState, useEffect } from 'react';
-
-import { useSnackbar } from 'notistack';
 import { format, formatDuration, intervalToDuration, Locale, parse, sub } from 'date-fns';
 import { enUS } from 'date-fns/locale';
-
-import { getPodLogs } from '../../../api/api';
-
-import { GridRenderCellParams } from '@mui/x-data-grid-pro';
-import { Link } from '@mui/material';
-import { DownloadRounded } from '@mui/icons-material';
 
 const shortDurationLocale: Locale = {
     formatDistance(token, count) {
@@ -73,31 +64,4 @@ export const formatStartingDate = (initialDate: string) => {
     }
 
     return format(parsedDate, 'yyyy/MM/dd HH:mm:ss');
-};
-
-export const DownloadLogs = (params: GridRenderCellParams) => {
-    const [data, setData] = useState({ logs: '' });
-
-    const { enqueueSnackbar } = useSnackbar();
-
-    useEffect(() => {
-        getPodLogs(params.row.application, params.row.podName, params.row.containerName).then(setData);
-    }, [params.row.application, params.row.podName, params.row.containerName]);
-
-    const logsBlob = new Blob([data.logs], { type: 'text/plain' });
-    const containerImage = params.row.image.replace(':', '/');
-
-    const handleNotification = () => {
-        enqueueSnackbar(`'${containerImage}' logs have been downloaded.`);
-    };
-
-    return (
-        <Link
-            href={URL.createObjectURL(logsBlob)}
-            download={`${containerImage}.log`}
-            onClick={handleNotification}
-        >
-            <DownloadRounded fontSize='small' sx={{ color: 'text.primary' }} />
-        </Link>
-    );
 };
