+++ Source/SelfService/Web/microservice/helpers/stringHelpers.ts
